requirements.txt
flask>=3.1.1,<4.0.0
flask-sqlalchemy>=3.1.1,<4.0.0
flask-restx>=1.3.0,<2.0.0
flask-cors>=6.0.1,<7.0.0
flask-jwt-extended>=4.7.1,<5.0.0
python-dotenv>=1.1.0,<2.0.0
alembic>=1.16.2,<2.0.0
apscheduler>=3.11.0,<4.0.0
weasyprint>=65.1,<66.0
pillow>=11.2.1,<12.0.0
pdfminer-six>=20250506,<20250507
flask-migrate>=4.1.0,<5.0.0
marshmallow>=4.0.0,<5.0.0
marshmallow-sqlalchemy>=1.4.2,<2.0.0
requests>=2.32.4,<3.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9 